Prevent submitting the section form with validation errors.
Under some circumstances like chaging a top-level select with conditional subforms with required fields, the section's Submit button will be enabled even if not all required fields have been filled. In these cases Formik's isValid value is true but validationSchema.isValidSync(formikValues) is false.
This happens because Formik for some reason is invoking the validation callback with stale data instead of the latest one, thus wrongly setting the valid boolean.

For now we add a workaround where we manually validate the values before submitting the form/going to the next section

This appears to be a Formik bug, possibly related to https://github.com/formium/formik/issues/1950.

See MOV-622